requirements.txt
streamlit
langchain
chromadb
sentence-transformers
python-dotenv
PyPDF2
tqdm
groq

